added more security

+++ routes/hisaab-route.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const isLoggedIn = require("../middleware/isLoggedIn");
+const isLoggedIn = require("../middleware/isLoggedIn-middleware");
 const userModel = require("../models/user-model");
 const hisaabModel = require("../models/hisaab-model");
 
@@ -45,7 +45,8 @@ router.post("/create", isLoggedIn, async (req, res) => {
 router.get("/view/:id", isLoggedIn, async (req, res) => {
   try {
     const hisaab = await hisaabModel.findOne({ _id: req.params.id });
-    res.render("view", { hisaab });
+    const user = await userModel.findOne({ _id: req.user.userId });
+    res.render("view", { hisaab, user });
   } catch (error) {
     res.redirect("/error");
   }
@@ -56,14 +57,18 @@ router.get("/delete/:id", isLoggedIn, async (req, res) => {
     const user = await userModel
       .findOne({ _id: req.user.userId })
       .populate("hisaab");
-    await hisaabModel.findOneAndDelete({ _id: req.params.id });
-    user.hisaab = user.hisaab.filter((h) => {
-      return h._id.toString() !== req.params.id;
-    });
-
-    await user.save();
-
-    res.redirect("/profile");
+    const hisaab = await hisaabModel.findOne({ _id: req.params.id });
+    console.log(hisaab.user, user._id);
+    if (user._id.toString() === hisaab.user.toString()) {
+      await hisaabModel.findOneAndDelete({ _id: req.params.id });
+      user.hisaab = user.hisaab.filter((h) => {
+        return h._id.toString() !== req.params.id;
+      });
+      await user.save();
+      res.redirect("/profile");
+    } else {
+      res.redirect("/profile");
+    }
   } catch (error) {
     res.redirect("/error");
   }
@@ -79,6 +84,7 @@ router.get("/edit/:id", isLoggedIn, async (req, res) => {
     res.redirect("/error");
   }
 });
+// 667330dd774305f94618b685
 
 router.post("/edit/:id", isLoggedIn, async (req, res) => {
   try {
@@ -88,11 +94,17 @@ router.post("/edit/:id", isLoggedIn, async (req, res) => {
     shareable = shareable ? shareable : false;
     editPermission = editPermission ? editPermission : false;
 
-    const hisaab = await hisaabModel.findOneAndUpdate(
-      { _id: req.params.id },
-      { title, description, encrypted, shareable, passcode, editPermission }
-    );
-    res.redirect("/profile");
+    const checkHisaab = await hisaabModel.findOne({ _id: req.params.id });
+
+    if (req.user.userId === checkHisaab.user.toString()) {
+      const hisaab = await hisaabModel.findOneAndUpdate(
+        { _id: req.params.id },
+        { title, description, encrypted, shareable, passcode, editPermission }
+      );
+      res.redirect("/profile");
+    } else {
+      res.send("ni manega");
+    }
   } catch (error) {
     res.redirect("/error");
   }
@@ -110,9 +122,10 @@ router.get("/view/passcode/:id", isLoggedIn, async (req, res) => {
 router.post("/view/passcode/:id", isLoggedIn, async (req, res) => {
   try {
     const hisaab = await hisaabModel.findOne({ _id: req.params.id });
+    const user = await userModel.findOne({ _id: req.user.userId });
 
     if (hisaab.passcode === Number(req.body.passcode)) {
-      res.render("view", { hisaab });
+      res.render("view", { hisaab, user });
     } else {
       req.flash("error", "Passcode incorrect");
       res.redirect(`/hisaab/view/passcode/${req.params.id}`);
